refactor: split commands.py into modular package structure
Split the 877-line commands.py file into focused modules:

**New Structure:**
- `commands/base.py` (99 lines) - Base classes, CommandSpec, CommandCategory
- `commands/registry.py` (222 lines) - CommandRegistry, CommandFactory
- `commands/implementations/` - Command classes by category:
  - `system.py` (177 lines) - HelpCommand, ClearCommand, UpdateCommand, RefreshConfigCommand
  - `debug.py` (189 lines) - YoloCommand, ThoughtsCommand, IterationsCommand, FixCommand, etc.
  - `development.py` (77 lines) - BranchCommand, InitCommand
  - `model.py` (61 lines) - ModelCommand
  - `conversation.py` (115 lines) - CompactCommand

**Benefits:**
- All files now under 500-line limit (largest: 222 lines)
- Clear separation of concerns by command category
- Easier to find and modify specific commands
- Better maintainability and test organization
- Full backward compatibility via package __init__.py

**Validation:**
- All 22 characterization tests pass
- No behavioral changes
- Proper import structure maintained
- Clean linting (black, isort, flake8)

Fixes #30
